feat: Add role field to User model
The code changes add a new field, role, to the User model. This field is of type String and has an enum constraint with values 'admin' and 'user'. The default value for the role field is set to 'user'. This feature enhances the functionality of the User model by allowing users to have different roles within the system.